src/requirements.txt
requests==2.28.1
pydantic==1.9.1
python-dotenv~=0.17.0
pycryptodome==3.15.0
